Added Auth Put and Delete Todo

+++ routers/todos.py
@@ -34,14 +34,14 @@ class TodoRequest(BaseModel):
 @router.get('/', status_code=status.HTTP_200_OK)
 async def read_all(user: user_dependency,db: db_dependency):
     if user is None:
-        raise HTTPException(status_code=401, detail='Authentiction Failed')
+        raise HTTPException(status_code=401, detail='Authentication Failed')
     return db.query(Todos).filter(Todos.owner == user.get('id')).all()
 
 
 @router.get('/todo/{todo_id}', status_code=status.HTTP_200_OK)
 async def get_todo(user: user_dependency,db: db_dependency, todo_id: int = Path(gt=0)):
     if user is None:
-        raise HTTPException(status_code=401, detail='Authentiction Failed')
+        raise HTTPException(status_code=401, detail='Authentication Failed')
     todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get('id')).first()
     if todo_model is not None:
         return todo_model
@@ -51,17 +51,20 @@ async def get_todo(user: user_dependency,db: db_dependency, todo_id: int = Path(
 @router.post("/todo", status_code=status.HTTP_200_OK)
 async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
     if user is None:
-        raise HTTPException(status_code=401, detail='Authentiction Failed')
+        raise HTTPException(status_code=401, detail='Authentication Failed')
     todo_model = Todos(**todo_request.dict(), owner=user.get('id'))
     db.add(todo_model)
     db.commit()
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def update_todo(db: db_dependency,
+async def update_todo(user: user_dependency,
+                      db: db_dependency,
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
     todo_model.title = todo_request.title
@@ -73,9 +76,11 @@ async def update_todo(db: db_dependency,
 
 
 @router.delete("/todo/{todo_id}")
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get('id')).delete()
     db.commit()
